audio: moved OpusInMemoryLoader into own file
